MAINT: random: Rename legacy distributions file.
In numpy/random/src/legacy/, rename distributions-boxmuller.c to
legacy-distributions.c.  This file holds all the legacy distributions,
so the new name is more accurate.
